requirements.txt
# requirements.txt

Flask
Flask-Bootstrap
Werkzeug
tensorflow
opencv-python
numpy
Pillow
